+++ hadoop-hdfs-project/hadoop-hdfs/CHANGES.txt
@@ -67,6 +67,9 @@ Release 2.8.0 - UNRELEASED
     HDFS-8073. Split BlockPlacementPolicyDefault.chooseTarget(..) so it
     can be easily overrided. (Walter Su via vinayakumarb)
 
+    HDFS-8080. Separate JSON related routines used by WebHdfsFileSystem to a
+    package local class. (wheat9)
+
   OPTIMIZATIONS
 
     HDFS-8026. Trace FSOutputSummer#writeChecksumChunks rather than
