+++ hadoop-hdfs-project/hadoop-hdfs/src/main/java/org/apache/hadoop/hdfs/web/JsonUtil.java
@@ -20,52 +20,37 @@ package org.apache.hadoop.hdfs.web;
 import org.apache.hadoop.fs.ContentSummary;
 import org.apache.hadoop.fs.FileChecksum;
 import org.apache.hadoop.fs.FileStatus;
-import org.apache.hadoop.fs.MD5MD5CRC32CastagnoliFileChecksum;
 import org.apache.hadoop.fs.MD5MD5CRC32FileChecksum;
-import org.apache.hadoop.fs.MD5MD5CRC32GzipFileChecksum;
 import org.apache.hadoop.fs.XAttr;
 import org.apache.hadoop.fs.XAttrCodec;
+import org.apache.hadoop.hdfs.XAttrHelper;
 import org.apache.hadoop.fs.permission.AclEntry;
 import org.apache.hadoop.fs.permission.AclStatus;
 import org.apache.hadoop.fs.permission.FsPermission;
-import org.apache.hadoop.hdfs.DFSUtil;
-import org.apache.hadoop.hdfs.XAttrHelper;
 import org.apache.hadoop.hdfs.protocol.DatanodeInfo;
-import org.apache.hadoop.hdfs.protocol.DatanodeInfo.AdminStates;
 import org.apache.hadoop.hdfs.protocol.ExtendedBlock;
 import org.apache.hadoop.hdfs.protocol.HdfsFileStatus;
 import org.apache.hadoop.hdfs.protocol.LocatedBlock;
 import org.apache.hadoop.hdfs.protocol.LocatedBlocks;
-import org.apache.hadoop.hdfs.security.token.block.BlockTokenIdentifier;
-import org.apache.hadoop.hdfs.security.token.delegation.DelegationTokenIdentifier;
-import org.apache.hadoop.hdfs.server.namenode.INode;
-import org.apache.hadoop.hdfs.server.blockmanagement.BlockStoragePolicySuite;
 import org.apache.hadoop.ipc.RemoteException;
 import org.apache.hadoop.security.token.Token;
 import org.apache.hadoop.security.token.TokenIdentifier;
-import org.apache.hadoop.util.DataChecksum;
 import org.apache.hadoop.util.StringUtils;
 import org.codehaus.jackson.map.ObjectMapper;
-import org.codehaus.jackson.map.ObjectReader;
 
 import com.google.common.collect.Maps;
 
-import java.io.ByteArrayInputStream;
-import java.io.DataInputStream;
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.TreeMap;
-import org.apache.hadoop.hdfs.protocol.FsAclPermission;
 
 /**
  * JSON Utilities
  */
 public class JsonUtil {
   private static final Object[] EMPTY_OBJECT_ARRAY = {};
-  private static final DatanodeInfo[] EMPTY_DATANODE_INFO_ARRAY = {};
 
   /**
    * Convert a token object to a Json string.
@@ -86,40 +71,6 @@ public class JsonUtil {
     return m;
   }
 
-  /**
-   * Convert a Json map to a Token.
-   */
-  public static Token<? extends TokenIdentifier> toToken(final Map<?, ?> m)
-      throws IOException {
-    if (m == null) {
-      return null;
-    }
-
-    final Token<DelegationTokenIdentifier> token =
-        new Token<>();
-    token.decodeFromUrlString((String) m.get("urlString"));
-    return token;
-  }
-
-  /**
-   * Convert a Json map to a Token of DelegationTokenIdentifier.
-   */
-  @SuppressWarnings("unchecked")
-  public static Token<DelegationTokenIdentifier> toDelegationToken(
-      final Map<?, ?> json) throws IOException {
-    final Map<?, ?> m = (Map<?, ?>) json.get(Token.class.getSimpleName());
-    return (Token<DelegationTokenIdentifier>) toToken(m);
-  }
-
-  /**
-   * Convert a Json map to a Token of BlockTokenIdentifier.
-   */
-  @SuppressWarnings("unchecked")
-  private static Token<BlockTokenIdentifier> toBlockToken(final Map<?, ?> m)
-      throws IOException {
-    return (Token<BlockTokenIdentifier>) toToken(m);
-  }
-
   /** Convert an exception object to a Json string. */
   public static String toJsonString(final Exception e) {
     final Map<String, Object> m = new TreeMap<>();
@@ -127,17 +78,6 @@ public class JsonUtil {
     m.put("message", e.getMessage());
     m.put("javaClassName", e.getClass().getName());
     return toJsonString(RemoteException.class, m);
-  }
-
-  /**
-   * Convert a Json map to a RemoteException.
-   */
-  public static RemoteException toRemoteException(final Map<?, ?> json) {
-    final Map<?, ?> m =
-        (Map<?, ?>) json.get(RemoteException.class.getSimpleName());
-    final String message = (String) m.get("message");
-    final String javaClassName = (String) m.get("javaClassName");
-    return new RemoteException(javaClassName, message);
   }
 
   private static String toJsonString(final Class<?> clazz, final Object value) {
@@ -165,27 +105,7 @@ public class JsonUtil {
     return String.format("%o", permission.toShort());
   }
 
-  /**
-   * Convert a string to a FsPermission object.
-   */
-  private static FsPermission toFsPermission(final String s, Boolean aclBit) {
-    FsPermission perm = new FsPermission(Short.parseShort(s, 8));
-    return (aclBit != null && aclBit) ? new FsAclPermission(perm) : perm;
-  }
-
-  static enum PathType {
-    FILE,
-    DIRECTORY,
-    SYMLINK;
-    
-    static PathType valueOf(HdfsFileStatus status) {
-      return status.isDir() ? DIRECTORY : status.isSymlink() ? SYMLINK : FILE;
-    }
-  }
-
-  /**
-   * Convert a HdfsFileStatus object to a Json string.
-   */
+  /** Convert a HdfsFileStatus object to a Json string. */
   public static String toJsonString(final HdfsFileStatus status,
       boolean includeType) throws IOException {
     if (status == null) {
@@ -193,7 +113,7 @@ public class JsonUtil {
     }
     final Map<String, Object> m = new TreeMap<>();
     m.put("pathSuffix", status.getLocalName());
-    m.put("type", PathType.valueOf(status));
+    m.put("type", WebHdfsConstants.PathType.valueOf(status));
     if (status.isSymlink()) {
       m.put("symlink", status.getSymlink());
     }
@@ -223,49 +143,8 @@ public class JsonUtil {
     return null;
   }
 
-  /**
-   * Convert a Json map to a HdfsFileStatus object.
-   */
-  public static HdfsFileStatus toFileStatus(final Map<?, ?> json,
-      boolean includesType) {
-    if (json == null) {
-      return null;
-    }
-
-    final Map<?, ?> m =
-        includesType ? (Map<?, ?>) json.get(FileStatus.class.getSimpleName()) :
-            json;
-    final String localName = (String) m.get("pathSuffix");
-    final PathType type = PathType.valueOf((String) m.get("type"));
-    final byte[] symlink = type != PathType.SYMLINK ? null :
-        DFSUtil.string2Bytes((String) m.get("symlink"));
-
-    final long len = ((Number) m.get("length")).longValue();
-    final String owner = (String) m.get("owner");
-    final String group = (String) m.get("group");
-    final FsPermission permission = toFsPermission((String) m.get("permission"),
-      (Boolean)m.get("aclBit"));
-    final long aTime = ((Number) m.get("accessTime")).longValue();
-    final long mTime = ((Number) m.get("modificationTime")).longValue();
-    final long blockSize = ((Number) m.get("blockSize")).longValue();
-    final short replication = ((Number) m.get("replication")).shortValue();
-    final boolean isFileStoredInDB = m.containsKey("isFileStoredInDB")? ((Boolean) m.get("isFileStoredInDB") ): false;
-    final long fileId = m.containsKey("fileId") ?
-        ((Number) m.get("fileId")).longValue() : INode.ROOT_PARENT_ID;
-    final int childrenNum = getInt(m, "childrenNum", -1);
-    final byte storagePolicy = m.containsKey("storagePolicy") ?
-        (byte) ((Number) m.get("storagePolicy")).longValue() :
-        BlockStoragePolicySuite.ID_UNSPECIFIED;
-    return new HdfsFileStatus(len, type == PathType.DIRECTORY, replication,
-        blockSize, mTime, aTime, permission, owner, group, symlink,
-        DFSUtil.string2Bytes(localName), fileId, childrenNum, isFileStoredInDB, storagePolicy);
-  }
-
-  /**
-   * Convert an ExtendedBlock to a Json map.
-   */
-  private static Map<String, Object> toJsonMap(
-      final ExtendedBlock extendedblock) {
+  /** Convert an ExtendedBlock to a Json map. */
+  private static Map<String, Object> toJsonMap(final ExtendedBlock extendedblock) {
     if (extendedblock == null) {
       return null;
     }
@@ -278,27 +157,8 @@ public class JsonUtil {
     return m;
   }
 
-  /**
-   * Convert a Json map to an ExtendedBlock object.
-   */
-  private static ExtendedBlock toExtendedBlock(final Map<?, ?> m) {
-    if (m == null) {
-      return null;
-    }
-    
-    final String blockPoolId = (String)m.get("blockPoolId");
-    final long blockId = ((Number) m.get("blockId")).longValue();
-    final long numBytes = ((Number) m.get("numBytes")).longValue();
-    final long generationStamp =
-        ((Number) m.get("generationStamp")).longValue();
-    return new ExtendedBlock(blockPoolId, blockId, numBytes, generationStamp);
-  }
-  
-  /**
-   * Convert a DatanodeInfo to a Json map.
-   */
-  static Map<String, Object> toJsonMap(
-      final DatanodeInfo datanodeinfo) {
+  /** Convert a DatanodeInfo to a Json map. */
+  static Map<String, Object> toJsonMap(final DatanodeInfo datanodeinfo) {
     if (datanodeinfo == null) {
       return null;
     }
@@ -327,101 +187,8 @@ public class JsonUtil {
     m.put("adminState", datanodeinfo.getAdminState().name());
     return m;
   }
- 
-  private static int getInt(Map<?, ?> m, String key, final int defaultValue) {
-    Object value = m.get(key);
-    if (value == null) {
-      return defaultValue;
-    }
-    return ((Number) value).intValue();
-  }
 
-  private static long getLong(Map<?, ?> m, String key, final long defaultValue) {
-    Object value = m.get(key);
-    if (value == null) {
-      return defaultValue;
-    }
-    return ((Number) value).longValue();
-  }
-  
-  private static String getString(Map<?, ?> m, String key,
-      final String defaultValue) {
-    Object value = m.get(key);
-    if (value == null) {
-      return defaultValue;
-    }
-    return (String) value;
-  }
-
-  static List<?> getList(Map<?, ?> m, String key) {
-    Object list = m.get(key);
-    if (list instanceof List<?>) {
-      return (List<?>) list;
-    } else {
-      return null;
-    }
-  }
-
-  /** Convert a Json map to an DatanodeInfo object. */
-  static DatanodeInfo toDatanodeInfo(final Map<?, ?> m)
-      throws IOException {
-    if (m == null) {
-      return null;
-    }
-    
-    // ipAddr and xferPort are the critical fields for accessing data.
-    // If any one of the two is missing, an exception needs to be thrown.
-
-    // Handle the case of old servers (1.x, 0.23.x) sending 'name' instead
-    //  of ipAddr and xferPort.
-    String ipAddr = getString(m, "ipAddr", null);
-    int xferPort = getInt(m, "xferPort", -1);
-    if (ipAddr == null) {
-      String name = getString(m, "name", null);
-      if (name != null) {
-        int colonIdx = name.indexOf(':');
-        if (colonIdx > 0) {
-          ipAddr = name.substring(0, colonIdx);
-          xferPort = Integer.parseInt(name.substring(colonIdx +1));
-        } else {
-          throw new IOException(
-              "Invalid value in server response: name=[" + name + "]");
-        }
-      } else {
-        throw new IOException(
-            "Missing both 'ipAddr' and 'name' in server response.");
-      }
-      // ipAddr is non-null & non-empty string at this point.
-    }
-
-    // Check the validity of xferPort.
-    if (xferPort == -1) {
-      throw new IOException(
-          "Invalid or missing 'xferPort' in server response.");
-    }
-
-    return new DatanodeInfo(ipAddr,
-        (String) m.get("hostName"), (String) m.get("storageID"),
-        xferPort,
-        ((Number) m.get("infoPort")).intValue(),
-        getInt(m, "infoSecurePort", 0),
-        ((Number) m.get("ipcPort")).intValue(),
-
-        getLong(m, "capacity", 0l),
-        getLong(m, "dfsUsed", 0l),
-        getLong(m, "remaining", 0l),
-        getLong(m, "blockPoolUsed", 0l),
-        getLong(m, "cacheCapacity", 0l),
-        getLong(m, "cacheUsed", 0l),
-        getLong(m, "lastUpdate", 0l),
-        getInt(m, "xceiverCount", 0),
-        getString(m, "networkLocation", ""),
-        AdminStates.valueOf(getString(m, "adminState", "NORMAL")));
-  }
-
-  /**
-   * Convert a DatanodeInfo[] to a Json array.
-   */
+  /** Convert a DatanodeInfo[] to a Json array. */
   private static Object[] toJsonArray(final DatanodeInfo[] array) {
     if (array == null) {
       return null;
@@ -436,28 +203,9 @@ public class JsonUtil {
     }
   }
 
-  /** Convert an Object[] to a DatanodeInfo[]. */
-  private static DatanodeInfo[] toDatanodeInfoArray(final List<?> objects)
-      throws IOException {
-    if (objects == null) {
-      return null;
-    } else if (objects.isEmpty()) {
-      return EMPTY_DATANODE_INFO_ARRAY;
-    } else {
-      final DatanodeInfo[] array = new DatanodeInfo[objects.size()];
-      int i = 0;
-      for (Object object : objects) {
-        array[i++] = toDatanodeInfo((Map<?, ?>) object);
-      }
-      return array;
-    }
-  }
-  
-  /**
-   * Convert a LocatedBlock to a Json map.
-   */
-  private static Map<String, Object> toJsonMap(final LocatedBlock locatedblock)
-      throws IOException {
+  /** Convert a LocatedBlock to a Json map. */
+  private static Map<String, Object> toJsonMap(final LocatedBlock locatedblock
+      ) throws IOException {
     if (locatedblock == null) {
       return null;
     }
@@ -472,34 +220,9 @@ public class JsonUtil {
     return m;
   }
 
-  /**
-   * Convert a Json map to LocatedBlock.
-   */
-  private static LocatedBlock toLocatedBlock(final Map<?, ?> m)
-      throws IOException {
-    if (m == null) {
-      return null;
-    }
-
-    final ExtendedBlock b = toExtendedBlock((Map<?, ?>)m.get("block"));
-    final DatanodeInfo[] locations = toDatanodeInfoArray(
-        getList(m, "locations"));
-    final long startOffset = ((Number) m.get("startOffset")).longValue();
-    final boolean isCorrupt = (Boolean)m.get("isCorrupt");
-    final DatanodeInfo[] cachedLocations = toDatanodeInfoArray(
-        getList(m, "cachedLocations"));
-
-    final LocatedBlock locatedblock = new LocatedBlock(b, locations,
-        null, null, startOffset, isCorrupt, cachedLocations);
-    locatedblock.setBlockToken(toBlockToken((Map<?, ?>)m.get("blockToken")));
-    return locatedblock;
-  }
-
-  /**
-   * Convert a LocatedBlock[] to a Json array.
-   */
-  private static Object[] toJsonArray(final List<LocatedBlock> array)
-      throws IOException {
+  /** Convert a LocatedBlock[] to a Json array. */
+  private static Object[] toJsonArray(final List<LocatedBlock> array
+      ) throws IOException {
     if (array == null) {
       return null;
     } else if (array.size() == 0) {
@@ -513,27 +236,9 @@ public class JsonUtil {
     }
   }
 
-  /** Convert an List of Object to a List of LocatedBlock. */
-  private static List<LocatedBlock> toLocatedBlockList(
-      final List<?> objects) throws IOException {
-    if (objects == null) {
-      return null;
-    } else if (objects.isEmpty()) {
-      return Collections.emptyList();
-    } else {
-      final List<LocatedBlock> list = new ArrayList<>(objects.size());
-      for (Object object : objects) {
-        list.add(toLocatedBlock((Map<?, ?>) object));
-      }
-      return list;
-    }
-  }
-
-  /**
-   * Convert LocatedBlocks to a Json string.
-   */
-  public static String toJsonString(final LocatedBlocks locatedblocks)
-      throws IOException {
+  /** Convert LocatedBlocks to a Json string. */
+  public static String toJsonString(final LocatedBlocks locatedblocks
+      ) throws IOException {
     if (locatedblocks == null) {
       return null;
     }
@@ -548,30 +253,7 @@ public class JsonUtil {
     return toJsonString(LocatedBlocks.class, m);
   }
 
-  /**
-   * Convert a Json map to LocatedBlock.
-   */
-  public static LocatedBlocks toLocatedBlocks(final Map<?, ?> json)
-      throws IOException {
-    if (json == null) {
-      return null;
-    }
-
-    final Map<?, ?> m = (Map<?, ?>)json.get(LocatedBlocks.class.getSimpleName());
-    final long fileLength = ((Number) m.get("fileLength")).longValue();
-    final boolean isUnderConstruction = (Boolean)m.get("isUnderConstruction");
-    final List<LocatedBlock> locatedBlocks = toLocatedBlockList(
-        getList(m, "locatedBlocks"));
-    final LocatedBlock lastLocatedBlock = toLocatedBlock(
-        (Map<?, ?>)m.get("lastLocatedBlock"));
-    final boolean isLastBlockComplete = (Boolean)m.get("isLastBlockComplete");
-    return new LocatedBlocks(fileLength, isUnderConstruction, locatedBlocks,
-        lastLocatedBlock, isLastBlockComplete);
-  }
-
-  /**
-   * Convert a ContentSummary to a Json string.
-   */
+  /** Convert a ContentSummary to a Json string. */
   public static String toJsonString(final ContentSummary contentsummary) {
     if (contentsummary == null) {
       return null;
@@ -587,30 +269,7 @@ public class JsonUtil {
     return toJsonString(ContentSummary.class, m);
   }
 
-  /**
-   * Convert a Json map to a ContentSummary.
-   */
-  public static ContentSummary toContentSummary(final Map<?, ?> json) {
-    if (json == null) {
-      return null;
-    }
-
-    final Map<?, ?> m = (Map<?, ?>)json.get(ContentSummary.class.getSimpleName());
-    final long length = ((Number) m.get("length")).longValue();
-    final long fileCount = ((Number) m.get("fileCount")).longValue();
-    final long directoryCount = ((Number) m.get("directoryCount")).longValue();
-    final long quota = ((Number) m.get("quota")).longValue();
-    final long spaceConsumed = ((Number) m.get("spaceConsumed")).longValue();
-    final long spaceQuota = ((Number) m.get("spaceQuota")).longValue();
-
-    return new ContentSummary.Builder().length(length).fileCount(fileCount).
-        directoryCount(directoryCount).quota(quota).spaceConsumed(spaceConsumed).
-        spaceQuota(spaceQuota).build();
-  }
-
-  /**
-   * Convert a MD5MD5CRC32FileChecksum to a Json string.
-   */
+  /** Convert a MD5MD5CRC32FileChecksum to a Json string. */
   public static String toJsonString(final MD5MD5CRC32FileChecksum checksum) {
     if (checksum == null) {
       return null;
@@ -623,54 +282,6 @@ public class JsonUtil {
     return toJsonString(FileChecksum.class, m);
   }
 
-  /**
-   * Convert a Json map to a MD5MD5CRC32FileChecksum.
-   */
-  public static MD5MD5CRC32FileChecksum toMD5MD5CRC32FileChecksum(
-      final Map<?, ?> json) throws IOException {
-    if (json == null) {
-      return null;
-    }
-
-    final Map<?, ?> m = (Map<?, ?>)json.get(FileChecksum.class.getSimpleName());
-    final String algorithm = (String)m.get("algorithm");
-    final int length = ((Number) m.get("length")).intValue();
-    final byte[] bytes = StringUtils.hexStringToByte((String)m.get("bytes"));
-
-    final DataInputStream in =
-        new DataInputStream(new ByteArrayInputStream(bytes));
-    final DataChecksum.Type crcType =
-        MD5MD5CRC32FileChecksum.getCrcTypeFromAlgorithmName(algorithm);
-    final MD5MD5CRC32FileChecksum checksum;
-
-    // Recreate what DFSClient would have returned.
-    switch (crcType) {
-      case CRC32:
-        checksum = new MD5MD5CRC32GzipFileChecksum();
-        break;
-      case CRC32C:
-        checksum = new MD5MD5CRC32CastagnoliFileChecksum();
-        break;
-      default:
-        throw new IOException("Unknown algorithm: " + algorithm);
-    }
-    checksum.readFields(in);
-
-    //check algorithm name
-    if (!checksum.getAlgorithmName().equals(algorithm)) {
-      throw new IOException(
-          "Algorithm not matched. Expected " + algorithm + ", Received " +
-              checksum.getAlgorithmName());
-    }
-    //check length
-    if (length != checksum.getLength()) {
-      throw new IOException(
-          "Length not matched: length=" + length + ", checksum.getLength()=" +
-              checksum.getLength());
-    }
-
-    return checksum;
-  }
   /** Convert a AclStatus object to a Json string. */
   public static String toJsonString(final AclStatus status) {
     if (status == null) {
@@ -708,35 +319,6 @@ public class JsonUtil {
     } catch (IOException ignored) {
     }
     return null;
-  }
-
-  /** Convert a Json map to a AclStatus object. */
-  public static AclStatus toAclStatus(final Map<?, ?> json) {
-    if (json == null) {
-      return null;
-    }
-
-    final Map<?, ?> m = (Map<?, ?>) json.get(AclStatus.class.getSimpleName());
-
-    AclStatus.Builder aclStatusBuilder = new AclStatus.Builder();
-    aclStatusBuilder.owner((String) m.get("owner"));
-    aclStatusBuilder.group((String) m.get("group"));
-    aclStatusBuilder.stickyBit((Boolean) m.get("stickyBit"));
-    String permString = (String) m.get("permission");
-    if (permString != null) {
-      final FsPermission permission = toFsPermission(permString,
-          (Boolean) m.get("aclBit"));
-      aclStatusBuilder.setPermission(permission);
-    }
-    final List<?> entries = (List<?>) m.get("entries");
-
-    List<AclEntry> aclEntryList = new ArrayList<AclEntry>();
-    for (Object entry : entries) {
-      AclEntry aclEntry = AclEntry.parseAclEntry((String) entry, true);
-      aclEntryList.add(aclEntry);
-    }
-    aclStatusBuilder.addEntries(aclEntryList);
-    return aclStatusBuilder.build();
   }
   
   public static String toJsonString(final XAttr xAttr,
@@ -805,6 +387,15 @@ public class JsonUtil {
     return XAttrHelper.buildXAttr(name, decodeXAttrValue(value));
   }
   
+  static List<?> getList(Map<?, ?> m, String key) {
+    Object list = m.get(key);
+    if (list instanceof List<?>) {
+      return (List<?>) list;
+    } else {
+      return null;
+    }
+  }
+    
   public static Map<String, byte[]> toXAttrs(final Map<?, ?> json)
       throws IOException {
     if (json == null) {
