+++ hadoop-hdfs-project/hadoop-hdfs/src/test/java/org/apache/hadoop/hdfs/web/TestJsonUtil.java
@@ -76,7 +76,7 @@ public class TestJsonUtil {
     System.out.println("json    = " + json.replace(",", ",\n  "));
     ObjectReader reader = new ObjectMapper().reader(Map.class);
     final HdfsFileStatus s2 =
-        JsonUtil.toFileStatus((Map<?, ?>) reader.readValue(json), true);
+        JsonUtilClient.toFileStatus((Map<?, ?>) reader.readValue(json), true);
     final FileStatus fs2 = toFileStatus(s2, parent);
     System.out.println("s2      = " + s2);
     System.out.println("fs2     = " + fs2);
@@ -105,7 +105,7 @@ public class TestJsonUtil {
     response.put("cacheCapacity", 123l);
     response.put("cacheUsed", 321l);
     
-    JsonUtil.toDatanodeInfo(response);
+    JsonUtilClient.toDatanodeInfo(response);
   }
   
   @Test
@@ -131,7 +131,7 @@ public class TestJsonUtil {
     response.put("cacheCapacity", 123l);
     response.put("cacheUsed", 321l);
 
-    DatanodeInfo di = JsonUtil.toDatanodeInfo(response);
+    DatanodeInfo di = JsonUtilClient.toDatanodeInfo(response);
     Assert.assertEquals(name, di.getXferAddr());
 
     // The encoded result should contain name, ipAddr and xferPort.
@@ -178,7 +178,7 @@ public class TestJsonUtil {
     aclStatusBuilder.stickyBit(false);
 
     Assert.assertEquals("Should be equal", aclStatusBuilder.build(),
-        JsonUtil.toAclStatus(json));
+        JsonUtilClient.toAclStatus(json));
   }
   
   @Test
@@ -245,7 +245,7 @@ public class TestJsonUtil {
   
   private void checkDecodeFailure(Map<String, Object> map) {
     try {
-      JsonUtil.toDatanodeInfo(map);
+      JsonUtilClient.toDatanodeInfo(map);
       Assert.fail("Exception not thrown against bad input.");
     } catch (Exception e) {
       // expected
